+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/web/PurchaseOrderLineController.java
@@ -28,7 +28,6 @@ import com.axelor.apps.base.service.tax.AccountManagementService;
 import com.axelor.apps.base.service.tax.FiscalPositionService;
 import com.axelor.apps.purchase.db.PurchaseOrder;
 import com.axelor.apps.purchase.db.PurchaseOrderLine;
-import com.axelor.apps.purchase.exception.IExceptionMessage;
 import com.axelor.apps.purchase.service.PurchaseOrderLineService;
 import com.axelor.db.mapper.Mapper;
 import com.axelor.exception.AxelorException;
@@ -41,6 +40,8 @@ import com.axelor.rpc.Context;
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import org.apache.commons.lang3.StringUtils;
+
 import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Optional;
@@ -108,11 +109,9 @@ public class PurchaseOrderLineController {
       String productCode =
           purchaseOrderLineService.getProductSupplierInfos(purchaseOrder, purchaseOrderLine)[1];
 
-      if (price == null || inTaxPrice == null || productName == null || productCode == null) {
-        response.setFlash(I18n.get(IExceptionMessage.PURCHASE_ORDER_LINE_NO_SUPPLIER_CATALOG));
-        resetProductInformation(response);
-        return;
-      }
+      // Do not interrupt process if product is not in catalog. Purchase price will use
+      // its default value if configuration says so (handled in get*TaxUnitPrice).
+      if (StringUtils.isBlank(productName)) productName = product.getName();
 
       response.setValue("unit", purchaseOrderLineService.getPurchaseUnit(purchaseOrderLine));
       response.setValue("qty", purchaseOrderLineService.getQty(purchaseOrder, purchaseOrderLine));